maybe try using u64 by default

+++ day5/src/lib.rs
@@ -1,19 +1,19 @@
 #[derive(Debug)]
 pub struct Range {
-    destination_start: u32,
-    source_start: u32,
-    length: u32,
+    destination_start: u64,
+    source_start: u64,
+    length: u64,
 }
 
 impl Range {
-    pub fn new(destination_start: u32, source_start: u32, length: u32) -> Range {
+    pub fn new(destination_start: u64, source_start: u64, length: u64) -> Range {
         Range {
             destination_start,
             source_start,
             length,
         }
     }
-    pub fn map(&self, input: u32) -> Option<u32> {
+    pub fn map(&self, input: u64) -> Option<u64> {
         if input < self.source_start || input >= self.source_start + self.length {
             return None;
         }
